adq patch - refactor economy : move declaration of class interfaceWeeklyOptimization in a new file
