+++ src/solver/cmake/solver.cmake
@@ -51,6 +51,7 @@ set(RTESOLVER_OPT
 		optimisation/opt_nombre_min_groupes_demarres_couts_demarrage.cpp
 		optimisation/opt_export_structure.h
 		optimisation/opt_export_structure.cpp
+		optimisation/base_weekly_optimization.h
 		optimisation/adequacy_patch.h
 		optimisation/adequacy_patch.cpp
 		optimisation/adequacy_patch_csr/solve_problem.cpp
